requirements.txt
# Core
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
pydantic-settings==2.7.1

# Microsoft Agent Framework (public preview — successor to Semantic Kernel + AutoGen)
agent-framework --pre

# Azure Identity
azure-identity==1.19.0

# Azure OpenAI (retained for embedding endpoint used by AI Search)
openai==1.58.1

# Azure AI Search
azure-search-documents==11.6.0b8

# Azure Monitor / Application Insights
azure-monitor-opentelemetry==1.6.4
opentelemetry-api==1.29.0
opentelemetry-sdk==1.29.0

# Redis (optional — for multi-instance session store)
redis[hiredis]==5.2.1

# Utilities
python-dotenv==1.0.1

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.28.1
requests==2.32.3
